Move viewmode buttons out of the sidebar
