+++ app/(home)/home/layout.tsx
@@ -1,14 +1,7 @@
-import Raffle from "components/raffle/Raffle";
-
 export default function HomeLayout({
 	children, // will be a page or nested layout
 }: {
 	children: React.ReactNode;
 }) {
-	return (
-		<>
-			<Raffle />
-			{children}
-		</>
-	);
+	return <>{children}</>;
 }
